refactor(firmware): SSOT-driven dispatcher sizing, HIL not-implemented simplification, pulse timeout overflow fix
- Generate FERQON_MAX_COMMAND_ID, FERQON_COMMAND_ID_COUNT, and
  FERQON_MAX_DRIVERS from protocol/ssot/commands.json so table sizes
  track the SSOT automatically.
- Add firmware_implemented flag to HIL method entries; emit
  hil_not_implemented directly from the generated dispatch table.
- Remove redundant hil_adc_read/hil_adc_expect/hil_pulse_measure stubs.
- Fix pulse timeout multiplication to avoid 16-bit overflow on AVR.

+++ src/pulse.cpp
@@ -25,7 +25,7 @@ static bool pulse_measure_handler(uint8_t seq, uint8_t cmd_id,
     /* Measure pulse width using pulseIn.
      * BLOCKING: no other commands or heartbeats processed during this. */
     /* pulseIn(pin, state, timeout) - timeout is in microseconds */
-    unsigned long pulse_us = ferqon_hal_pulse_in(pin, 1, timeout_ms * 1000);
+    unsigned long pulse_us = ferqon_hal_pulse_in(pin, 1, (unsigned long)timeout_ms * 1000UL);
 
     if (pulse_us == 0) {
         /* Timeout or no pulse */
